# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"AI QA Automation Platform","description":"AI-powered QA test management and browser automation platform","version":"0.2.0"},"paths":{"/projects/":{"get":{"tags":["Projects"],"summary":"Get Projects","operationId":"get_projects_projects__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ProjectResponse"},"type":"array","title":"Response Get Projects Projects  Get"}}}}}},"post":{"tags":["Projects"],"summary":"Create Project","operationId":"create_project_projects__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/projects/{project_id}":{"get":{"tags":["Projects"],"summary":"Get Project","operationId":"get_project_projects__project_id__get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/projects/{project_id}/applications/":{"post":{"tags":["Applications"],"summary":"Create Application","operationId":"create_application_projects__project_id__applications__post","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Applications"],"summary":"Get Applications","operationId":"get_applications_projects__project_id__applications__get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ApplicationResponse"},"title":"Response Get Applications Projects  Project Id  Applications  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/projects/{project_id}/applications/{application_id}":{"get":{"tags":["Applications"],"summary":"Get Application","operationId":"get_application_projects__project_id__applications__application_id__get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}},{"name":"application_id","in":"path","required":true,"schema":{"type":"integer","title":"Application Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/projects/{project_id}/applications/{application_id}/discover":{"post":{"tags":["Discovery"],"summary":"Discover","operationId":"discover_projects__project_id__applications__application_id__discover_post","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}},{"name":"application_id","in":"path","required":true,"schema":{"type":"integer","title":"Application Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/projects/{project_id}/applications/{application_id}/test-cases/":{"post":{"tags":["Test Cases"],"summary":"Create Test Case","operationId":"create_test_case_projects__project_id__applications__application_id__test_cases__post","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}},{"name":"application_id","in":"path","required":true,"schema":{"type":"integer","title":"Application Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TestCaseCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TestCaseResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Test Cases"],"summary":"Get Test Cases","operationId":"get_test_cases_projects__project_id__applications__application_id__test_cases__get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}},{"name":"application_id","in":"path","required":true,"schema":{"type":"integer","title":"Application Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TestCaseResponse"},"title":"Response Get Test Cases Projects  Project Id  Applications  Application Id  Test Cases  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/projects/{project_id}/applications/{application_id}/test-cases/{test_case_id}":{"get":{"tags":["Test Cases"],"summary":"Get Test Case","operationId":"get_test_case_projects__project_id__applications__application_id__test_cases__test_case_id__get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}},{"name":"application_id","in":"path","required":true,"schema":{"type":"integer","title":"Application Id"}},{"name":"test_case_id","in":"path","required":true,"schema":{"type":"integer","title":"Test Case Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TestCaseResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/projects/{project_id}/applications/{application_id}/test-cases/{test_case_id}/steps/":{"post":{"tags":["Test Steps"],"summary":"Create Test Step","operationId":"create_test_step_projects__project_id__applications__application_id__test_cases__test_case_id__steps__post","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}},{"name":"application_id","in":"path","required":true,"schema":{"type":"integer","title":"Application Id"}},{"name":"test_case_id","in":"path","required":true,"schema":{"type":"integer","title":"Test Case Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TestStepCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TestStepResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Test Steps"],"summary":"Get Test Steps","operationId":"get_test_steps_projects__project_id__applications__application_id__test_cases__test_case_id__steps__get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}},{"name":"application_id","in":"path","required":true,"schema":{"type":"integer","title":"Application Id"}},{"name":"test_case_id","in":"path","required":true,"schema":{"type":"integer","title":"Test Case Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TestStepResponse"},"title":"Response Get Test Steps Projects  Project Id  Applications  Application Id  Test Cases  Test Case Id  Steps  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/projects/{project_id}/applications/{application_id}/test-cases/{test_case_id}/runs/":{"post":{"tags":["Test Runs"],"summary":"Run Test Case","operationId":"run_test_case_projects__project_id__applications__application_id__test_cases__test_case_id__runs__post","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}},{"name":"application_id","in":"path","required":true,"schema":{"type":"integer","title":"Application Id"}},{"name":"test_case_id","in":"path","required":true,"schema":{"type":"integer","title":"Test Case Id"}}],"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TestRunResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Test Runs"],"summary":"Get Runs","operationId":"get_runs_projects__project_id__applications__application_id__test_cases__test_case_id__runs__get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}},{"name":"application_id","in":"path","required":true,"schema":{"type":"integer","title":"Application Id"}},{"name":"test_case_id","in":"path","required":true,"schema":{"type":"integer","title":"Test Case Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TestRunResponse"},"title":"Response Get Runs Projects  Project Id  Applications  Application Id  Test Cases  Test Case Id  Runs  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/projects/{project_id}/applications/{application_id}/test-cases/{test_case_id}/runs/{run_id}/steps":{"get":{"tags":["Test Runs"],"summary":"Get Run Steps","operationId":"get_run_steps_projects__project_id__applications__application_id__test_cases__test_case_id__runs__run_id__steps_get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}},{"name":"application_id","in":"path","required":true,"schema":{"type":"integer","title":"Application Id"}},{"name":"test_case_id","in":"path","required":true,"schema":{"type":"integer","title":"Test Case Id"}},{"name":"run_id","in":"path","required":true,"schema":{"type":"integer","title":"Run Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TestRunStepResponse"},"title":"Response Get Run Steps Projects  Project Id  Applications  Application Id  Test Cases  Test Case Id  Runs  Run Id  Steps Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/projects/{project_id}/applications/{application_id}/test-cases/{test_case_id}/runs/{run_id}/steps/{step_id}/screenshot":{"get":{"tags":["Test Runs"],"summary":"Get Step Screenshot","operationId":"get_step_screenshot_projects__project_id__applications__application_id__test_cases__test_case_id__runs__run_id__steps__step_id__screenshot_get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}},{"name":"application_id","in":"path","required":true,"schema":{"type":"integer","title":"Application Id"}},{"name":"test_case_id","in":"path","required":true,"schema":{"type":"integer","title":"Test Case Id"}},{"name":"run_id","in":"path","required":true,"schema":{"type":"integer","title":"Run Id"}},{"name":"step_id","in":"path","required":true,"schema":{"type":"integer","title":"Step Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/projects/{project_id}/applications/{application_id}/ai/generate-test":{"post":{"tags":["AI Test Generation"],"summary":"Generate Test","operationId":"generate_test_projects__project_id__applications__application_id__ai_generate_test_post","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}},{"name":"application_id","in":"path","required":true,"schema":{"type":"integer","title":"Application Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AITestGenerationRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AITestGenerationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RootResponse"}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}}},"components":{"schemas":{"AITestGenerationRequest":{"properties":{"requirement":{"type":"string","maxLength":5000,"minLength":10,"title":"Requirement"}},"type":"object","required":["requirement"],"title":"AITestGenerationRequest"},"AITestGenerationResponse":{"properties":{"test_case":{"$ref":"#/components/schemas/GeneratedTestCase"}},"type":"object","required":["test_case"],"title":"AITestGenerationResponse"},"ApplicationCreate":{"properties":{"name":{"type":"string","maxLength":200,"minLength":1,"title":"Name"},"url":{"type":"string","title":"Url"}},"type":"object","required":["name","url"],"title":"ApplicationCreate"},"ApplicationResponse":{"properties":{"id":{"type":"integer","title":"Id"},"project_id":{"type":"integer","title":"Project Id"},"name":{"type":"string","title":"Name"},"url":{"type":"string","title":"Url"}},"type":"object","required":["id","project_id","name","url"],"title":"ApplicationResponse"},"GeneratedTestCase":{"properties":{"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description"},"priority":{"type":"string","title":"Priority"},"steps":{"items":{"$ref":"#/components/schemas/GeneratedTestStep"},"type":"array","title":"Steps"}},"type":"object","required":["name","description","priority","steps"],"title":"GeneratedTestCase"},"GeneratedTestStep":{"properties":{"step_number":{"type":"integer","title":"Step Number"},"action":{"type":"string","title":"Action"},"description":{"type":"string","title":"Description"},"target":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target"},"value":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Value"},"expected_result":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Expected Result"}},"type":"object","required":["step_number","action","description"],"title":"GeneratedTestStep"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status"}},"type":"object","required":["status"],"title":"HealthResponse"},"ProjectCreate":{"properties":{"name":{"type":"string","maxLength":200,"minLength":1,"title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["name"],"title":"ProjectCreate"},"ProjectResponse":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["id","name"],"title":"ProjectResponse"},"RootResponse":{"properties":{"message":{"type":"string","title":"Message"},"status":{"type":"string","title":"Status"},"version":{"type":"string","title":"Version"}},"type":"object","required":["message","status","version"],"title":"RootResponse"},"TestCaseCreate":{"properties":{"name":{"type":"string","maxLength":300,"minLength":1,"title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"priority":{"type":"string","title":"Priority","default":"medium"}},"type":"object","required":["name"],"title":"TestCaseCreate"},"TestCaseResponse":{"properties":{"id":{"type":"integer","title":"Id"},"application_id":{"type":"integer","title":"Application Id"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"priority":{"type":"string","title":"Priority"},"status":{"type":"string","title":"Status"}},"type":"object","required":["id","application_id","name","priority","status"],"title":"TestCaseResponse"},"TestRunResponse":{"properties":{"id":{"type":"integer","title":"Id"},"test_case_id":{"type":"integer","title":"Test Case Id"},"status":{"type":"string","title":"Status"},"started_at":{"type":"string","format":"date-time","title":"Started At"},"finished_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Finished At"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"}},"type":"object","required":["id","test_case_id","status","started_at"],"title":"TestRunResponse"},"TestRunStepResponse":{"properties":{"id":{"type":"integer","title":"Id"},"test_run_id":{"type":"integer","title":"Test Run Id"},"test_step_id":{"type":"integer","title":"Test Step Id"},"status":{"type":"string","title":"Status"},"message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Message"},"started_at":{"type":"string","format":"date-time","title":"Started At"},"finished_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Finished At"},"duration_ms":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Duration Ms"},"screenshot_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Screenshot Path"}},"type":"object","required":["id","test_run_id","test_step_id","status","started_at"],"title":"TestRunStepResponse"},"TestStepCreate":{"properties":{"step_number":{"type":"integer","minimum":1.0,"title":"Step Number"},"action":{"type":"string","maxLength":50,"minLength":1,"title":"Action"},"description":{"type":"string","title":"Description"},"target":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target"},"value":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Value"},"expected_result":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Expected Result"}},"type":"object","required":["step_number","action","description"],"title":"TestStepCreate"},"TestStepResponse":{"properties":{"id":{"type":"integer","title":"Id"},"test_case_id":{"type":"integer","title":"Test Case Id"},"step_number":{"type":"integer","title":"Step Number"},"action":{"type":"string","title":"Action"},"description":{"type":"string","title":"Description"},"target":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target"},"value":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Value"},"expected_result":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Expected Result"}},"type":"object","required":["id","test_case_id","step_number","action","description"],"title":"TestStepResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}